test: 구간 생성 테스트 추가
