--- src/test/java/nextstep/subway/unit/LineServiceTest.java
@@ -1,33 +1,0 @@
-package nextstep.subway.unit;
-
-import nextstep.subway.application.LineService;
-import nextstep.subway.infrastructure.LineRepository;
-import nextstep.subway.infrastructure.StationRepository;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
-
-@SpringBootTest
-@Transactional
-public class LineServiceTest {
-    @Autowired
-    private StationRepository stationRepository;
-    @Autowired
-    private LineRepository lineRepository;
-
-    @Autowired
-    private LineService lineService;
-
-    @Test
-    void addSection() {
-        // given
-        // stationRepository와 lineRepository를 활용하여 초기값 셋팅
-
-        // when
-        // lineService.addSection 호출
-
-        // then
-        // line.getSections 메서드를 통해 검증
-    }
-}
